seperated home page from the main.dart file
