+++ lib/cat_settings.dart
@@ -1,7 +1,7 @@
 /* Cat Logger Cat Settings Page.*/
 import 'package:flutter/material.dart';
 
-import 'main.dart';
+import 'home.dart';
 import 'user_settings.dart';
 
 class MySettingsPage extends StatefulWidget {
